requirements.txt
Django
django_compressor
django-libsass
django-haystack
